Multithread tests

+++ perftest/maps/maptests.cpp
@@ -184,7 +184,7 @@ private:
 
 typedef adapter::hash_map<slow_int_type, slow_int_type, dummy_hash<slow_int_type::type> > generic_hash_map_type;
 typedef adapter::hash_map<long long, long long> ikey_hash_map_type;
-typedef adapter::hash_map<int, int> ipair_hash_map_type;
+typedef adapter::hash_map<long long, int> ipair_hash_map_type;
 typedef adapter::hash_trie<int, int, 16> hash_trie_type;
 typedef adapter::stdmap<int, int, false> map_type;
 typedef adapter::stdmap<int, int, true> unorderd_map_type;
